Removed '-Wno-double-promotion' from the build flags. This necessitated fixing float -> double promotions in several places in the code. In most cases, double was never being utilized anyway, so there was a slight performance hit for casting when not required.

+++ src/common/tv_filters/NTSCFilter.cxx
@@ -20,8 +20,8 @@
 
 #include "NTSCFilter.hxx"
 
-constexpr double scaleFrom100(double x) { return (x/50.0) - 1.0;     }
-constexpr uInt32 scaleTo100(double x)   { return uInt32(50*(x+1.0)); }
+constexpr float scaleFrom100(float x) { return (x/50.f) - 1.f;     }
+constexpr uInt32 scaleTo100(float x)  { return uInt32(50*(x+1.f)); }
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 NTSCFilter::NTSCFilter()
